2025/08/18 - 001
Rearranged the HoldNote and HoldNoteEditor, now in the HoldNote there's no more start_time and end_time variable, now there's a _duration var.
The start_time is the _current_time (from Note class) and the end_time it's the start_time plus the duration, the set_end_time changes the value of the duration by this logic: duration = end_time - start_time.

Bug Fix:
Solved the problem in the NoteInfo of the NoteHold displaying wrong values of the time and going above the song duration.

TODO: Make impossible for the user to create a HoldNote with the end_time under the start_time.